+++ infobip-openapi-mcp-core/src/main/java/com/infobip/openapi/mcp/openapi/tool/ToolHandler.java
@@ -7,7 +7,9 @@ import com.infobip.openapi.mcp.enricher.ApiRequestEnricherChain;
 import com.infobip.openapi.mcp.error.ErrorModelWriter;
 import com.infobip.openapi.mcp.infrastructure.metrics.MetricService;
 import com.infobip.openapi.mcp.openapi.schema.DecomposedRequestData;
+import com.infobip.openapi.mcp.progress.ProgressUpdateProvider;
 import io.modelcontextprotocol.spec.McpSchema;
+import java.util.Objects;
 import java.util.Optional;
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
@@ -65,6 +67,8 @@ public class ToolHandler {
     private final ApiRequestEnricherChain enricherChain;
     private final MetricService metricService;
     private final CredentialProvider credentialProvider;
+    private final ProgressUpdateProvider progressUpdateProvider;
+    private final NotificationTicker ticker;
 
     public ToolHandler(
             RestClient restClient,
@@ -72,13 +76,41 @@ public class ToolHandler {
             OpenApiMcpProperties properties,
             ApiRequestEnricherChain enricherChain,
             MetricService metricService,
-            CredentialProvider credentialProvider) {
+            CredentialProvider credentialProvider,
+            ProgressUpdateProvider progressUpdateProvider) {
+        this(
+                restClient,
+                errorModelWriter,
+                properties,
+                enricherChain,
+                metricService,
+                credentialProvider,
+                progressUpdateProvider,
+                () -> Thread.sleep(properties.progressNotificationsInterval().toMillis()));
+    }
+
+    /**
+     * Internal constructor used in tests only. Allows for mocking the passage of
+     * time in tests by overriding the default Thread.sleep implementation of
+     * the ticker.
+     */
+    ToolHandler(
+            RestClient restClient,
+            ErrorModelWriter errorModelWriter,
+            OpenApiMcpProperties properties,
+            ApiRequestEnricherChain enricherChain,
+            MetricService metricService,
+            CredentialProvider credentialProvider,
+            ProgressUpdateProvider progressUpdateProvider,
+            NotificationTicker ticker) {
         this.restClient = restClient;
         this.errorModelWriter = errorModelWriter;
         this.properties = properties;
         this.enricherChain = enricherChain;
         this.metricService = metricService;
         this.credentialProvider = credentialProvider;
+        this.progressUpdateProvider = progressUpdateProvider;
+        this.ticker = ticker;
         this.serializationCorrector = new JsonDoubleSerializationCorrector();
     }
 
@@ -276,7 +308,73 @@ public class ToolHandler {
         enrichedSpec.headers(
                 headers -> headers.addIfAbsent(HttpHeaders.ACCEPT, PREFERRED_ACCEPT_MEDIA_TYPE.toString()));
 
-        return enrichedSpec.retrieve().toEntity(String.class);
+        if (!properties.progressNotificationsEnabled()
+                || context.progressNotification() == null
+                || context.callToolRequest() == null
+                || context.callToolRequest().progressToken() == null) {
+            return enrichedSpec.retrieve().toEntity(String.class);
+        }
+
+        Double total;
+        try {
+            total = progressUpdateProvider.total(context);
+        } catch (Throwable throwable) {
+            LOGGER.warn(
+                    "Failed to initialise progress notifications because: {} \n"
+                            + "No progress notifications will be sent.",
+                    throwable.getMessage(),
+                    throwable);
+            return enrichedSpec.retrieve().toEntity(String.class);
+        }
+
+        var progressNotifThread = Thread.ofVirtual().start(notificationLoop(total, context));
+
+        try {
+            var result = enrichedSpec.retrieve().toEntity(String.class);
+            progressNotifThread.interrupt();
+            try {
+                progressNotifThread.join();
+            } catch (InterruptedException joinInterruptedException) {
+                throw new RuntimeException(joinInterruptedException);
+            }
+            return result;
+        } catch (RuntimeException apiCallFail) {
+            progressNotifThread.interrupt();
+            try {
+                progressNotifThread.join();
+            } catch (InterruptedException ignored) {
+                Thread.currentThread().interrupt();
+            }
+            throw apiCallFail;
+        }
+    }
+
+    private Runnable notificationLoop(@Nullable Double total, McpRequestContext context) {
+        var progressNotificationConsumer = Objects.requireNonNull(context.progressNotification());
+        var progressToken = Objects.requireNonNull(context.callToolRequest()).progressToken();
+        return () -> {
+            var tick = 0L;
+            while (true) {
+                try {
+                    ticker.tick();
+                    var update = progressUpdateProvider.next(tick++, context);
+                    progressNotificationConsumer.accept(new McpSchema.ProgressNotification(
+                            progressToken, update.progress(), total, update.message()));
+                } catch (InterruptedException e) {
+                    // This is regular flow that happens when the main tool handler thread
+                    // interrupts the progress notif thread before writing the tool call
+                    // result back to MCP client.
+                    return;
+                } catch (Throwable throwable) {
+                    LOGGER.warn(
+                            "Failed to send progress notification because: {} \n"
+                                    + "No further progress notifications will be sent.",
+                            throwable.getMessage(),
+                            throwable);
+                    return;
+                }
+            }
+        };
     }
 
     /**