+++ infobip-openapi-mcp-core/src/test/java/com/infobip/openapi/mcp/openapi/schema/InputExampleComposerTest.java
@@ -706,6 +706,8 @@ class InputExampleComposerTest {
                 null,
                 null,
                 null,
+                null,
+                null,
                 new OpenApiMcpProperties.Tools(
                         null,
                         new OpenApiMcpProperties.Tools.Schema(parametersKey, requestBodyKey),