+++ infobip-openapi-mcp-core/src/test/java/com/infobip/openapi/mcp/openapi/tool/ToolRegistryTest.java
@@ -66,6 +66,8 @@ class ToolRegistryTest {
                 null,
                 null,
                 null,
+                null,
+                null,
                 new OpenApiMcpProperties.Tools(null, null, null, true, null, null, null),
                 null);
         toolRegistry = new ToolRegistry(
@@ -908,6 +910,8 @@ class ToolRegistryTest {
                 null,
                 null,
                 null,
+                null,
+                null,
                 new OpenApiMcpProperties.Tools(null, null, null, false, null, null, null),
                 null);
         toolRegistry = new ToolRegistry(
@@ -954,6 +958,8 @@ class ToolRegistryTest {
     void shouldUseSummaryAsDescriptionWhenFeatureDisabledAndOnlySummaryPresent() {
         // Given
         properties = new OpenApiMcpProperties(
+                null,
+                null,
                 null,
                 null,
                 null,
@@ -1299,6 +1305,8 @@ class ToolRegistryTest {
                 null,
                 null,
                 null,
+                null,
+                null,
                 new OpenApiMcpProperties.Tools(null, null, null, true, null, ExamplesMode.ALL, null),
                 null);
         inputExampleComposer = new InputExampleComposer(properties);
@@ -1356,6 +1364,8 @@ class ToolRegistryTest {
     void shouldNotAppendExampleBlockWhenFeatureDisabled() {
         // Given
         properties = new OpenApiMcpProperties(
+                null,
+                null,
                 null,
                 null,
                 null,
@@ -1452,6 +1462,8 @@ class ToolRegistryTest {
     void shouldReturnExampleBlockAsDescriptionWhenNoSummaryOrDescription() {
         // Given
         properties = new OpenApiMcpProperties(
+                null,
+                null,
                 null,
                 null,
                 null,
@@ -1561,6 +1573,8 @@ class ToolRegistryTest {
                 null,
                 null,
                 null,
+                null,
+                null,
                 new OpenApiMcpProperties.Tools(null, null, null, true, null, ExamplesMode.ALL, null),
                 null);
         inputExampleComposer = new InputExampleComposer(properties);
@@ -1629,6 +1643,8 @@ class ToolRegistryTest {
     void shouldAppendOnlyAnnotatedExamplesWhenModeIsAnnotated() {
         // Given
         properties = new OpenApiMcpProperties(
+                null,
+                null,
                 null,
                 null,
                 null,